Egy kis kiegeszites a view-ba es game logic-ba
minimalis a game logic valtozas, itt gyk csak kamu kod kerult be, hogy a
view-val rajzolhassak valamit.

+++ UniScrollShooter/View/Controls/Control.cs
@@ -104,7 +104,7 @@ namespace View.Controls
             color = Color.White;
         }
 
-        public virtual void Update(float elapsedTime)
+        public virtual void Update(double elapsedTime)
         {
         }
